requirements.txt
pyserial
keyboard
pygame